polish(ctx): setup ctx cancel in main
- dedups cancel logic
