+++ commands/walk.go
@@ -3,10 +3,7 @@ package commands
 import (
 	"context"
 	"errors"
-	"os"
-	"os/signal"
 	"strings"
-	"syscall"
 	"time"
 
 	"github.com/filecoin-project/sentinel-visor/schedule"
@@ -100,20 +97,6 @@ func walk(cctx *cli.Context) error {
 			strg = db
 		}
 	}
-	// Set up a context that is canceled when the command is interrupted
-	ctx, cancel := context.WithCancel(cctx.Context)
-	defer cancel()
-
-	// Set up a signal handler to cancel the context
-	go func() {
-		interrupt := make(chan os.Signal, 1)
-		signal.Notify(interrupt, syscall.SIGTERM, syscall.SIGINT)
-		select {
-		case <-interrupt:
-			cancel()
-		case <-ctx.Done():
-		}
-	}()
 
 	scheduler := schedule.NewScheduler(cctx.Duration("task-delay"))
 
@@ -136,7 +119,7 @@ func walk(cctx *cli.Context) error {
 	})
 
 	// Start the scheduler and wait for it to complete or to be cancelled.
-	err = scheduler.Run(ctx)
+	err = scheduler.Run(cctx.Context)
 	if !errors.Is(err, context.Canceled) {
 		return err
 	}
